requirements.txt
fastapi
asyncpg
uvicorn
datetime
requests
httpx
python-dotenv
gunicorn
Werkzeug==2.2.2
oauthlib
passlib
requests_oauthlib
python-multipart